# pkg/api/docs/swagger.json
{
    "components": {
        "schemas": {
...
            "api.ErrorResponse": {
                "properties": {
                    "error": {
                        "type": "string"
                    }
                },
                "required": [
                    "error"
                ],
                "type": "object"
            },
...
            "api.restoreRequest": {
                "properties": {
                    "as_new": {
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
            "engine.BackupInfo": {
                "properties": {
                    "consistent": {
                        "type": "boolean"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "disks": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "live": {
                        "type": "boolean"
                    },
                    "size_bytes": {
                        "type": "integer"
                    },
                    "timestamp": {
                        "type": "string"
                    },
                    "vm_name": {
                        "type": "string"
                    }
                },
                "required": [
                    "consistent",
                    "created_at",
                    "disks",
                    "live",
                    "size_bytes",
                    "timestamp",
                    "vm_name"
                ],
                "type": "object"
            },
...
            "engine.ScheduleParams": {
                "properties": {
                    "enabled": {
                        "type": "boolean"
                    },
                    "interval_hours": {
                        "type": "integer"
                    },
                    "keep_last": {
                        "type": "integer"
                    },
                    "max_age_days": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "jobs.Job": {
                "properties": {
                    "action": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "error": {
                        "type": "string"
                    },
                    "finished_at": {
                        "type": "string"
                    },
                    "id": {
                        "type": "string"
                    },
                    "label": {
                        "type": "string"
                    },
                    "logs": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "result": {
                        "type": "string"
                    },
                    "started_at": {
                        "type": "string"
                    },
                    "state": {
                        "$ref": "#/components/schemas/jobs.State"
                    },
                    "target": {
                        "type": "string"
                    }
                },
                "required": [
                    "action",
                    "created_at",
                    "id",
                    "label",
                    "logs",
                    "state",
                    "target"
                ],
                "type": "object"
            },
...
            "jobs.State": {
                "enum": [
                    "pending",
                    "running",
                    "succeeded",
                    "failed"
                ],
                "type": "string",
                "x-enum-varnames": [
                    "Pending",
                    "Running",
                    "Succeeded",
                    "Failed"
                ]
            },
...
            "types.BackupSchedule": {
                "properties": {
                    "enabled": {
                        "type": "boolean"
                    },
                    "interval_hours": {
                        "type": "integer"
                    },
                    "keep_last": {
                        "type": "integer"
                    },
                    "last_run_at": {
                        "type": "integer"
                    },
                    "max_age_days": {
                        "type": "integer"
                    },
                    "vm_id": {
                        "type": "string"
                    }
                },
                "required": [
                    "enabled",
                    "interval_hours",
                    "keep_last",
                    "last_run_at",
                    "max_age_days",
                    "vm_id"
                ],
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "BearerAuth": {
                "description": "Bearer JWT obtained from POST /api/login.",
                "in": "header",
                "name": "Authorization",
                "type": "apiKey"
            }
        }
    },
    "info": {
        "description": "VM lifecycle, networking, media and jobs API. Mutations returning 202 are asynchronous; inspect the returned job for completion.",
        "title": "Maco API",
        "version": "1.0"
    },
...
    "paths": {
...
        "/api/vms/{id}/backups": {
            "get": {
                "operationId": "listBackups",
                "parameters": [
                    {
                        "description": "id",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/engine.BackupInfo"
                                    },
                                    "type": "array"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "listBackups",
                "tags": [
                    "vms"
                ]
            },
            "post": {
                "operationId": "createBackup",
                "parameters": [
                    {
                        "description": "id",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/jobs.Job"
                                }
                            }
                        },
                        "description": "Accepted"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "503": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Service Unavailable"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "createBackup",
                "tags": [
                    "vms"
                ]
            }
        },
        "/api/vms/{id}/backups/schedule": {
            "get": {
                "operationId": "getBackupSchedule",
                "parameters": [
                    {
                        "description": "id",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/types.BackupSchedule"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "getBackupSchedule",
                "tags": [
                    "vms"
                ]
            },
            "put": {
                "operationId": "setBackupSchedule",
                "parameters": [
                    {
                        "description": "id",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/engine.ScheduleParams",
                                        "summary": "body",
                                        "description": "Request"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "Request",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/types.BackupSchedule"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "setBackupSchedule",
                "tags": [
                    "vms"
                ]
            }
        },
        "/api/vms/{id}/backups/{timestamp}": {
            "delete": {
                "operationId": "deleteBackup",
                "parameters": [
                    {
                        "description": "id",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "timestamp",
                        "in": "path",
                        "name": "timestamp",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/jobs.Job"
                                }
                            }
                        },
                        "description": "Accepted"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "503": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Service Unavailable"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "deleteBackup",
                "tags": [
                    "vms"
                ]
            }
        },
        "/api/vms/{id}/backups/{timestamp}/restore": {
            "post": {
                "operationId": "restoreBackup",
                "parameters": [
                    {
                        "description": "id",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "timestamp",
                        "in": "path",
                        "name": "timestamp",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/api.restoreRequest",
                                        "summary": "body",
                                        "description": "Request"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "Request",
                    "required": true
                },
                "responses": {
                    "202": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/jobs.Job"
                                }
                            }
                        },
                        "description": "Accepted"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "503": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Service Unavailable"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "restoreBackup",
                "tags": [
                    "vms"
                ]
            }
        },
...
    },
    "openapi": "3.1.0",
    "servers": [
        {
            "url": "/"
        }
    ]
}